Move Issue hooks into a separate module
